7zpp: don't truncate sizes larger than maxuint32
Archive members may be larger than 2**32-1. kpidSize is a VT_UI8 so per
MSDN, the correct field of the union is uhVal.

https://msdn.microsoft.com/en-us/library/windows/desktop/aa380072(v=vs.85).aspx

+++ endless/src/7zip-cpp/7zpp/SevenZipArchive.h
@@ -26,7 +26,7 @@ namespace SevenZip
 
 		virtual size_t GetNumberOfItems();
 		virtual std::vector<TString> GetItemsNames();
-		virtual std::vector<size_t>  GetOrigSizes();
+		virtual std::vector<UINT64>  GetOrigSizes();
 
 	protected:
 		bool m_ReadMetadata = false;
@@ -37,7 +37,7 @@ namespace SevenZip
 		CompressionLevelEnum m_compressionLevel;
 		size_t m_numberofitems = 0;
 		std::vector<TString> m_itemnames;
-		std::vector<size_t> m_origsizes;
+		std::vector<UINT64> m_origsizes;
 
 	private:
 		bool pri_GetNumberOfItems();
