+++ endless/src/7zip-cpp/7zpp/UsefulFunctions.h
@@ -29,7 +29,7 @@ namespace SevenZip
 
 		static bool UsefulFunctions::GetItemsNames(const SevenZipLibrary & library, const TString & archivePath,
 			CompressionFormatEnum & format, size_t & numberofitems, 
-			std::vector<TString> & itemnames, std::vector<size_t> & origsizes);
+			std::vector<TString> & itemnames, std::vector<UINT64> & origsizes);
 
 		static const TString EndingFromCompressionFormat(const CompressionFormatEnum& format);
 	};
